Updated with holidays and more color settings.
Foreground and background colors can now be specified for regular dates,
holidays and today's date. There are default parameters that can be
overridden by passing in an array of ConsoleColors.

I added four new functions to get a list of holidays throughout the
year. The holidays will show up on the calendar if you use the
ShowHolidays switch.