Added an ObservedHolidays switch
Using the ObservedHolidays switch will print the days on which holidays are
observed rather than the days they actually occur.

To find the observed day, I assumed that holidays on a Saturday are
observed on the preceding Friday and holidays on a Sunday are observed
on the succeeding Monday.

This heuristic seems to be in line with the Federal holiday schedule,
but not necessarily SIFMA's schedule. Case in point, SIFMA did not
recognize Friday, December 31 2010 as a holiday.